WIP added complete paste & melted with duplicate
composites and bindings not working
