+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/CopyPasteHelper.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
@@ -22,6 +23,7 @@ namespace UnityEngine.InputSystem.Editor
         };
 
         public static SerializedProperty lastNewElement;
+        private static SerializedProperty currentActionMap;
 
         public static void CopySelectedTreeViewItemsToClipboard(List<SerializedProperty> items, Type type)
         {
@@ -44,12 +46,7 @@ namespace UnityEngine.InputSystem.Editor
         private static void CopyItemData(SerializedProperty item, StringBuilder buffer)
         {
             buffer.Append(k_StartOfText);
-            // InputActionMaps have back-references to InputActionAssets. Make sure we ignore those.
-            buffer.Append(item.CopyToJson());
-
-            // if (!item.serializedDataIncludesChildren && item.hasChildrenIncludingHidden) //TODO: copying child data necessary?
-            //     foreach (var child in item.childrenIncludingHidden)
-            //         CopyItemData(child, buffer);
+            buffer.Append(item.CopyToJson(true));
         }
 
         public static bool HavePastableClipboardData(Type selectedType)
@@ -72,9 +69,10 @@ namespace UnityEngine.InputSystem.Editor
             return null;
         }
 
-        public static void PasteFromClipboard(int[] indicesToInsert, SerializedProperty arrayToInsertInto)
+        public static void PasteFromClipboard(int[] indicesToInsert, SerializedProperty arrayToInsertInto, SerializedProperty actionMap = null)
         {
             lastNewElement = null;
+            currentActionMap = actionMap;
             PasteData(EditorGUIUtility.systemCopyBuffer, indicesToInsert, arrayToInsertInto);
         }
 
@@ -82,17 +80,12 @@ namespace UnityEngine.InputSystem.Editor
         {
             if (!copyBufferString.StartsWith(k_CopyPasteMarker))
                 return;
-
-            ////REVIEW: filtering out children may remove the very item we need to get the right match for a copy block?
             PasteItems(copyBufferString, indicesToInsert, arrayToInsertInto);
         }
 
         private static void PasteItems(string copyBufferString, int[] indicesToInsert, SerializedProperty arrayToInsertInto, bool assignNewIDs = true, bool selectNewItems = true)
         {
-            var newItemPropertyPaths = new List<string>();
-
-            // Split buffer into transmissions and then into transmission blocks. Each transmission is an item subtree
-            // meant to be pasted as a whole and each transmission block is a single chunk of serialized data.
+            // Split buffer into transmissions and then into transmission blocks
             int indexOffset = 0;
             foreach (var transmission in copyBufferString.Substring(k_CopyPasteMarker.Length + k_TypeMarker[GetCopiedClipboardType()].Length)
                      .Split(new[] {k_EndOfTransmission}, StringSplitOptions.RemoveEmptyEntries))
@@ -105,12 +98,201 @@ namespace UnityEngine.InputSystem.Editor
 
         private static void PasteBlocks(string transmission, int indexToInsert, SerializedProperty arrayToInsertInto)
         {
-            var block = transmission.Substring(transmission.IndexOf(k_StartOfText) + 1);
-            var pastedProperty = InputActionSerializationHelpers.AddElement(arrayToInsertInto, "name", indexToInsert);
-            pastedProperty.RestoreFromJson(block);
-            InputActionSerializationHelpers.EnsureUniqueName(pastedProperty);
-            InputActionSerializationHelpers.AssignUniqueIDs(pastedProperty);
-            lastNewElement = pastedProperty;
+            var block = transmission.Substring(transmission.IndexOf(k_StartOfText, StringComparison.Ordinal) + 1);
+            var copiedType = GetCopiedClipboardType();
+            if (copiedType == typeof(InputActionMap))
+                PasteElement(arrayToInsertInto, block, indexToInsert);
+            else if (copiedType == typeof(InputAction))
+                PasteAction(currentActionMap, arrayToInsertInto, block, indexToInsert);
+            else
+            {
+                PasteBindingOrComposite(arrayToInsertInto, block, "", indexToInsert);
+            }
+        }
+
+        public static SerializedProperty DuplicateElement(SerializedProperty arrayProperty, SerializedProperty toDuplicate, string name, int index, bool changeName = true)
+        {
+            var json = toDuplicate.CopyToJson(true);
+            return PasteElement(arrayProperty, json, index, name, changeName);
+        }
+
+        private static SerializedProperty PasteElement(SerializedProperty arrayProperty, string json, int index, string name = "newElement",  bool changeName = true)
+        {
+            var duplicatedProperty = AddElement(arrayProperty, name, index);
+            duplicatedProperty.RestoreFromJson(json);
+            if (changeName)
+                EnsureUniqueName(duplicatedProperty);
+            AssignUniqueIDs(duplicatedProperty);
+            lastNewElement = duplicatedProperty;
+            return duplicatedProperty;
+        }
+
+        public static void DuplicateAction(SerializedProperty actionMap, SerializedProperty arrayProperty, SerializedProperty toDuplicate, string name)
+        {
+            var json = toDuplicate.CopyToJson(true);
+            PasteAction(actionMap, arrayProperty, json, toDuplicate.GetIndexOfArrayElement() + 1);
+        }
+
+        private static void PasteAction(SerializedProperty actionMap, SerializedProperty arrayProperty, string jsonToInsert, int indexToInsert)
+        {
+            var property = PasteElement(arrayProperty, jsonToInsert, indexToInsert, "", false);
+            var name = property.FindPropertyRelative("m_Name").stringValue;
+            EnsureUniqueName(property);
+            var newName = property.FindPropertyRelative("m_Name").stringValue;
+            var bindingsArray = actionMap.FindPropertyRelative(nameof(InputActionMap.m_Bindings));
+            var bindings = bindingsArray.Where(binding => binding.FindPropertyRelative("m_Action").stringValue.Equals(name)).ToList();
+            var index = bindings.Select(b => b.GetIndexOfArrayElement()).Max() + 1;
+            foreach (var binding in bindings)
+            {
+                var newIndex = PasteBindingAsPartOfAction(bindingsArray, binding, newName, index);
+                index = newIndex;
+            }
+        }
+
+        private static int PasteBindingAsPartOfAction(SerializedProperty arrayProperty, SerializedProperty toDuplicate, string newActionName, int index)
+        {
+            if (IsComposite(toDuplicate))
+            {
+                DuplicateComposite(arrayProperty, toDuplicate, PropertyName(toDuplicate), newActionName, index, out var newIndex, false);
+                return newIndex;
+            }
+            if (IsPartComposite(toDuplicate))
+                return index;
+            var duplicatedBinding = DuplicateElement(arrayProperty, toDuplicate, PropertyName(toDuplicate), index++, false);
+            duplicatedBinding.FindPropertyRelative("m_Action").stringValue = newActionName;
+            return index;
+        }
+
+        private static bool IsComposite(SerializedProperty property) => property.FindPropertyRelative("m_Flags").intValue == (int)InputBinding.Flags.Composite;
+        private static bool IsPartComposite(SerializedProperty property) => property.FindPropertyRelative("m_Flags").intValue == (int)InputBinding.Flags.PartOfComposite;
+        private static string PropertyName(SerializedProperty property) => property.FindPropertyRelative("m_Name").stringValue;
+
+        public static int DuplicateBinding(SerializedProperty arrayProperty, SerializedProperty toDuplicate, string newActionName, int index)
+        {
+            if (IsComposite(toDuplicate))
+                return DuplicateComposite(arrayProperty, toDuplicate, PropertyName(toDuplicate), newActionName, index, out _).GetIndexOfArrayElement();
+            var binding = DuplicateElement(arrayProperty, toDuplicate, newActionName, index, false);
+            return PasteBinding(binding, newActionName, index);
+        }
+
+        private static void PasteBindingOrComposite(SerializedProperty arrayProperty, string toDuplicate, string newActionName, int index)
+        {
+            var property = PasteElement(arrayProperty, toDuplicate, index++, "", false);
+            if (IsComposite(property))
+            {
+                PasteComposite(arrayProperty, property, PropertyName(property), newActionName, index);
+                return;
+            }
+            PasteBinding(property, newActionName, index);
+        }
+
+        private static int PasteBinding(SerializedProperty duplicatedBinding, string newActionName, int index)
+        {
+            duplicatedBinding.FindPropertyRelative("m_Action").stringValue = newActionName;
+            return index;
+        }
+
+        private static SerializedProperty PasteComposite(SerializedProperty bindingsArray, SerializedProperty duplicatedComposite, string name, string actionName, int index, bool increaseIndex = true)
+        {
+            index += InputActionSerializationHelpers.GetCompositePartCount(bindingsArray, duplicatedComposite.GetIndexOfArrayElement());
+            duplicatedComposite.FindPropertyRelative("m_Name").stringValue = name;
+            duplicatedComposite.FindPropertyRelative("m_Action").stringValue = actionName;
+            PasteBindingsForComposite(bindingsArray, duplicatedComposite, index, actionName);
+            return duplicatedComposite;
+        }
+
+        private static SerializedProperty DuplicateComposite(SerializedProperty bindingsArray, SerializedProperty compositeToDuplicate, string name, string actionName, int index, out int newIndex, bool increaseIndex = true)
+        {
+            if (increaseIndex)
+                index += InputActionSerializationHelpers.GetCompositePartCount(bindingsArray, compositeToDuplicate.GetIndexOfArrayElement());
+            var newComposite = DuplicateElement(bindingsArray, compositeToDuplicate, name, index++, false);
+            newComposite.FindPropertyRelative("m_Action").stringValue = actionName;
+
+            PasteBindingsForComposite(bindingsArray, compositeToDuplicate, index, actionName);
+            newIndex = index;
+            return newComposite;
+        }
+
+        private static void PasteBindingsForComposite(SerializedProperty bindingsArray, SerializedProperty compositeToDuplicate, int index, string actionName)
+        {
+            var bindings = GetBindingsForComposite(bindingsArray, compositeToDuplicate);
+            foreach (var binding in bindings)
+            {
+                var newBinding = DuplicateElement(bindingsArray, binding, binding.FindPropertyRelative("m_Name").stringValue, index++, false);
+                newBinding.FindPropertyRelative("m_Action").stringValue = actionName;
+            }
+        }
+
+        private static List<SerializedProperty> GetBindingsForComposite(SerializedProperty bindingsArray, SerializedProperty compositeToDuplicate)
+        {
+            var compositeBindings = new List<SerializedProperty>();
+            var compositeStartIndex = InputActionSerializationHelpers.GetCompositeStartIndex(bindingsArray, compositeToDuplicate.GetIndexOfArrayElement());
+            if (compositeStartIndex == -1)
+                return compositeBindings;
+
+            for (var i = compositeStartIndex + 1; i < bindingsArray.arraySize; ++i)
+            {
+                var bindingProperty = bindingsArray.GetArrayElementAtIndex(i);
+                var bindingFlags = (InputBinding.Flags)bindingProperty.FindPropertyRelative("m_Flags").intValue;
+                if ((bindingFlags & InputBinding.Flags.PartOfComposite) == 0)
+                    break;
+                compositeBindings.Add(bindingProperty);
+            }
+            return compositeBindings;
+        }
+
+        public static SerializedProperty AddElement(SerializedProperty arrayProperty, string name, int index = -1)
+        {
+            var uniqueName = InputActionSerializationHelpers.FindUniqueName(arrayProperty, name);
+            if (index < 0)
+                index = arrayProperty.arraySize;
+
+            arrayProperty.InsertArrayElementAtIndex(index);
+            var elementProperty = arrayProperty.GetArrayElementAtIndex(index);
+            elementProperty.ResetValuesToDefault();
+
+            elementProperty.FindPropertyRelative("m_Name").stringValue = uniqueName;
+            elementProperty.FindPropertyRelative("m_Id").stringValue = Guid.NewGuid().ToString();
+
+            return elementProperty;
+        }
+
+        public static void EnsureUniqueName(SerializedProperty arrayElement)
+        {
+            var arrayProperty = arrayElement.GetArrayPropertyFromElement();
+            var arrayIndexOfElement = arrayElement.GetIndexOfArrayElement();
+            var nameProperty = arrayElement.FindPropertyRelative("m_Name");
+            var baseName = nameProperty.stringValue;
+            nameProperty.stringValue = InputActionSerializationHelpers.FindUniqueName(arrayProperty, baseName, ignoreIndex: arrayIndexOfElement);
+        }
+
+        public static void AssignUniqueIDs(SerializedProperty element)
+        {
+            // Assign new ID to map.
+            AssignUniqueID(element);
+
+            //
+            foreach (var child in element.GetChildren())
+            {
+                if (!child.isArray)
+                    continue;
+
+                var fieldType = child.GetFieldType();
+                if (fieldType == typeof(InputBinding[]) || fieldType == typeof(InputAction[]) ||
+                    fieldType == typeof(InputActionMap))
+                {
+                    ////TODO: update bindings that refer to actions by {id}
+                    for (var i = 0; i < child.arraySize; ++i)
+                        using (var childElement = child.GetArrayElementAtIndex(i))
+                            AssignUniqueIDs(childElement);
+                }
+            }
+        }
+
+        private static void AssignUniqueID(SerializedProperty property)
+        {
+            var idProperty = property.FindPropertyRelative("m_Id");
+            idProperty.stringValue = Guid.NewGuid().ToString();
         }
     }
 }
